major refactoring + architecture change (whatsapp photo)

+++ fourmi.h
@@ -23,7 +23,7 @@ public:
     virtual void update();
     // export the entity to something that can be written in a file
     virtual std::vector<std::vector<std::string>> exportToString();
-    virtual std::vector<std::vector<int>> draw();
+    virtual std::vector<std::vector<double>> draw();
 };
 
 class Collector : public Fourmi {
@@ -38,7 +38,7 @@ public:
     // create the object and return its pointer + check that it dont overlap something
     static std::shared_ptr<Fourmi> importFromExtSaveCollector(
             std::vector<std::string> &inputBuffer, int index);
-    virtual std::vector<std::vector<int>> draw();
+    virtual std::vector<std::vector<double>> draw();
 };
 
 class Defensor : public Fourmi {
@@ -50,8 +50,7 @@ public:
     // create the object and return its pointer + check that it dont overlap something
     static std::shared_ptr<Fourmi> importFromExtSaveDefensor(
             std::vector<std::string> &inputBuffer, int index);
-    virtual std::vector<std::vector<int>> draw();
-
+    virtual std::vector<std::vector<double>> draw();
 };
 
 class Predator : public Fourmi {
@@ -64,7 +63,7 @@ public:
     static std::shared_ptr<Fourmi> importFromExtSavePredator(
             std::vector<std::string> &inputBuffer, int index);
 
-    virtual std::vector<std::vector<int>> draw();
+    virtual std::vector<std::vector<double>> draw();
 };
 
 class Generator : public Fourmi {
@@ -76,7 +75,7 @@ public:
     // create the object and return its pointer + check that it dont overlap something
     static std::shared_ptr<Fourmi> importFromExtSaveGenerator(
                                     std::vector<std::string> &inputBuffer, int index);
-    virtual std::vector<std::vector<int>> draw();
+    virtual std::vector<std::vector<double>> draw();
 };
 
 #endif //TCHANZ_FOURMI_H
